requirements.txt
python-telegram-bot>=20.7
aiohttp>=3.9
aiosqlite>=0.19
python-dotenv>=1.0
base58>=2.1
